# Microservice/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nerve IDP API
  version: 1.0.0
  description: |
    Internal Developer Platform API — contract-first specification.
    All backend services implement against this spec.
    All frontend React Query hooks are typed against these schemas.
  contact:
    name: Platform Engineering
  license:
    name: Private
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://nerve.internal/api/v1
    description: Production
...
paths:
...
  /services:
    get:
      tags: [catalog]
      summary: List all services with optional filters
      operationId: listServices
      parameters:
        - $ref: "#/components/parameters/TeamFilter"
        - $ref: "#/components/parameters/LanguageFilter"
        - $ref: "#/components/parameters/HealthFilter"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, maturity_score, error_budget, last_deploy_at, team]
            default: name
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated service list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [catalog]
      summary: Register a new service in the catalog
      operationId: registerService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRegistration"
      responses:
        "201":
          description: Service registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "409":
          description: Service name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /services/{serviceId}:
    get:
      tags: [catalog]
      summary: Get full service detail
      operationId: getService
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: Full service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [catalog]
      summary: Update service metadata
      operationId: updateService
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"

    delete:
      tags: [catalog]
      summary: Deregister a service (soft delete)
      operationId: deleteService
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "204":
          description: Service deregistered
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TeamFilter:
      name: team
      in: query
      schema:
        type: string
    LanguageFilter:
      name: language
      in: query
      schema:
        type: string
        enum: [python, go, typescript, rust, java]
    HealthFilter:
      name: health
      in: query
      schema:
        type: string
        enum: [healthy, degraded, frozen, unknown]
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient RBAC role for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Service:
      type: object
      required: [id, name, team, language, version, health_status, compliance_score,
                 maturity_score, error_budget_consumed, deploy_frozen, replica_count,
                 last_deploy_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: payment-service
        team:
          type: string
          example: commerce
        language:
          type: string
          enum: [python, go, typescript, rust, java]
        version:
          type: string
          example: v1.9.0
        repo_url:
          type: string
          format: uri
        health_status:
          type: string
          enum: [healthy, degraded, frozen, unknown]
        compliance_score:
          type: integer
          minimum: 0
          maximum: 100
        maturity_score:
          type: integer
          minimum: 0
          maximum: 100
        error_budget_consumed:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Percentage of 30-day error budget consumed
        deploy_frozen:
          type: boolean
        replica_count:
          type: integer
        last_deploy_at:
          type: string
          format: date-time
        template_version:
          type: string
          description: Scaffold template version used — used to flag services behind current standard
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServiceDetail:
      allOf:
        - $ref: "#/components/schemas/Service"
        - type: object
          properties:
            slo:
              $ref: "#/components/schemas/SloDefinition"
            error_budget:
              $ref: "#/components/schemas/ErrorBudget"
            security:
              $ref: "#/components/schemas/SecurityPostureSummary"
            maturity_breakdown:
              $ref: "#/components/schemas/MaturityScore"
            cost:
              $ref: "#/components/schemas/ServiceCostSummary"
            recent_deploys:
              type: array
              items:
                $ref: "#/components/schemas/DeployHistoryEntry"
            upstream_dependencies:
              type: array
              items:
                type: string
              description: Service IDs this service depends on
            downstream_dependents:
              type: array
              items:
                type: string
              description: Service IDs that depend on this service
    ServiceListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        summary:
          $ref: "#/components/schemas/CatalogSummary"
    CatalogSummary:
      type: object
      properties:
        total_services:
          type: integer
        healthy:
          type: integer
        degraded:
          type: integer
        frozen:
          type: integer
        avg_maturity_score:
          type: number
        critical_cves:
          type: integer
    ServiceRegistration:
      type: object
      required: [name, team, language, repo_url]
      properties:
        name:
          type: string
        team:
          type: string
        language:
          type: string
          enum: [python, go, typescript, rust, java]
        repo_url:
          type: string
          format: uri
        description:
          type: string
        upstream_dependencies:
          type: array
          items:
            type: string
            format: uuid
    ServiceUpdate:
      type: object
      properties:
        version:
          type: string
        replica_count:
          type: integer
        health_status:
          type: string
          enum: [healthy, degraded, frozen, unknown]
        description:
          type: string

    # ── SLO ──────────────────────────────────────
    SloDefinition:
      type: object
      required: [service_id, sli_type, target, window_days]
      properties:
        id:
          type: string
          format: uuid
        service_id:
          type: string
          format: uuid
        sli_type:
          type: string
          enum: [availability, latency, throughput, error_rate]
        target:
          type: number
          format: float
          description: Target percentage (e.g. 99.9 for 99.9% availability)
        window_days:
          type: integer
          default: 30
        latency_threshold_ms:
          type: integer
          description: Required when sli_type is latency
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    DeployHistoryEntry:
      type: object
      required: [deploy_id, version, environment, status, compliance_score, actor, deployed_at]
      properties:
        deploy_id:
          type: string
          format: uuid
        version:
          type: string
        environment:
          type: string
        status:
          type: string
          enum: [succeeded, failed, blocked, rolled_back]
        compliance_score:
          type: integer
        actor:
          type: string
        deployed_at:
          type: string
          format: date-time
...
    ErrorBudget:
      type: object
      required: [service_id, slo_target, budget_consumed, budget_remaining, burn_alerts, frozen]
      properties:
        service_id:
          type: string
          format: uuid
        slo_target:
          type: number
        budget_consumed:
          type: number
          description: Percentage 0-100
        budget_remaining:
          type: number
          description: Percentage 0-100
        frozen:
          type: boolean
        frozen_at:
          type: string
          format: date-time
        burn_alerts:
          type: array
          items:
            $ref: "#/components/schemas/BurnAlert"
        computed_at:
          type: string
          format: date-time
    BurnAlert:
      type: object
      required: [burn_rate, window, severity, firing]
      properties:
        burn_rate:
          type: number
          description: Current burn rate multiplier (1x = on track to exhaust in window)
        window:
          type: string
          enum: [1h, 6h, 1d, 3d]
        severity:
          type: string
          enum: [page, ticket, warning]
        firing:
          type: boolean
        time_to_exhaustion_hours:
          type: number
...
    MaturityScore:
      type: object
      required: [service_id, overall_score, pillars, computed_at]
      properties:
        service_id:
          type: string
          format: uuid
        overall_score:
          type: integer
          minimum: 0
          maximum: 100
        pillars:
          type: object
          properties:
            observability:
              $ref: "#/components/schemas/MaturityPillar"
            reliability:
              $ref: "#/components/schemas/MaturityPillar"
            security:
              $ref: "#/components/schemas/MaturityPillar"
            docs:
              $ref: "#/components/schemas/MaturityPillar"
            cost:
              $ref: "#/components/schemas/MaturityPillar"
            error_budget_health:
              $ref: "#/components/schemas/MaturityPillar"
        template_version:
          type: string
        template_behind_by:
          type: integer
          description: Number of major versions behind current scaffold template standard
        computed_at:
          type: string
          format: date-time
    MaturityPillar:
      type: object
      required: [score, weight, signals]
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        weight:
          type: number
        signals:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              passed:
                type: boolean
              detail:
                type: string
...
    ServiceCostSummary:
      type: object
      properties:
        current_month_usd:
          type: number
        anomaly_detected:
          type: boolean
...
    SecurityPostureSummary:
      type: object
      properties:
        score:
          type: integer
        critical_cves:
          type: integer
        last_scan_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
          format: uuid
security:
  - BearerAuth: []